requirements.txt
# Data Processing
pandas==2.0.3
numpy==1.24.3
scipy==1.10.1

# Machine Learning
scikit-learn==1.3.0
tensorflow==2.13.0
torch==2.0.1
xgboost==1.7.6
lightgbm==4.0.0

# Data Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Financial Analysis
yfinance==0.2.28
ta-lib==0.4.28
pandas-ta==0.3.14b

# Web Framework
flask==2.3.3
django==4.2.4
fastapi==0.101.1

# API
requests==2.31.0
aiohttp==3.8.5

# Database
sqlalchemy==2.0.19
pymongo==4.4.1
psycopg2-binary==2.9.7

# Development Tools
jupyter==1.0.0
ipython==8.14.0
pytest==7.4.0
black==23.7.0
flake8==6.1.0
isort==5.12.0

# Project Management
python-dotenv==1.0.0
pyyaml==6.0.1

# Deployment
docker==6.1.3
gunicorn==21.2.0
uvicorn==0.23.2

# Documentation
sphinx==7.1.2
mkdocs==1.5.2

# Version Control
gitpython==3.1.32

# Add to existing requirements.txt
fredapi==0.5.0
newsapi-python==0.2.7